requirements.txt
dj-database-url==0.3.0
Django==1.8.3
django-admin-sortable2==0.5.1
django-appconf==1.0.1
django-classy-tags==0.6.2
django-cms==3.1.2
django-embed-video==1.0.0
django-imagekit==3.2.6
django-polymorphic==0.7.1
django-realtime==1.1
django-redactoreditor==1.2.10
django-reversion==1.8.5
django-sekizai==0.8.2
Django-Select2==4.3.1
django-treebeard==3.0
django-versatileimagefield==1.0.3
djangocms-admin-style==0.2.7
djangocms-column==1.5
djangocms-file==0.1
djangocms-flash==0.2.0
djangocms-googlemap==0.2
djangocms-inherit==0.1
djangocms-installer==0.7.8
djangocms-link==1.6.2
djangocms-picture==0.1
djangocms-style==1.5
djangocms-teaser==0.1
djangocms-text-ckeditor==2.5.2
djangocms-video==0.1
html5lib==0.999999
pilkit==1.1.12
Pillow==2.8.0
pytz==2015.4
requests==2.7.0
six==1.9.0
tzlocal==1.2
wheel==0.24.0
